# Conflicts:
#	Server/mods/deathmatch/logic/CResourceChecker.Data.h

+++ Client/core/CCommands.cpp
@@ -102,7 +102,7 @@ bool CCommands::Execute(const char* szCommand, const char* szParametersIn, bool 
             // His line starts with '/'?
             if (*szParameters == '/')
             {
-                // Copy the characters after the slash to the 0 terminator to a seperate buffer
+                // Copy the characters after the slash to the 0 terminator to a separate buffer
                 char szBuffer[256];
                 strncpy(szBuffer, szParameters + 1, 256);
                 szBuffer[255] = 0;
@@ -176,7 +176,7 @@ bool CCommands::Execute(const char* szCommand, const char* szParametersIn, bool 
                 {
                     int fpsVal = 0;
                     CVARS_GET("fps_limit", fpsVal);
-                    CCore::GetSingleton().GetFPSLimiter()->SetUserDefinedFPS(fpsVal);
+                    CCore::GetSingleton().GetFPSLimiter()->SetUserDefinedFPS(static_cast<std::uint16_t>(fpsVal));
                 }
 
                 // HACK: Foul dirty hack to force vsync (Rework on #4427)
